added chapter adapter

+++ app/internal/domain/entity/chapter.go
@@ -8,7 +8,7 @@ type Chapter struct {
 	Name         string      `json:"name"`
 	Num          string      `json:"num,omitempty"`
 	RegulationID uint64      `json:"regulation_id,omitempty"`
-	OrderNum     uint64      `json:"order_num"`
+	OrderNum     uint32      `json:"order_num"`
 	Paragraphs   []Paragraph `json:"paragraphs,omitempty"`
 	UpdatedAt    *time.Time  `json:"updated_at,omitempty"`
 }
